refactor(WinningRequestView): 당첨 번호를 입력받는 뷰 클래스를 분리
