+++ src/main/java/lotto/view/constants/MessageType.java
@@ -3,7 +3,6 @@ package lotto.view.constants;
 import static lotto.view.constants.SymbolType.NEWLINE;
 
 public enum MessageType {
-    WINNING_NUMBERS_REQUEST_MESSAGE("당첨 번호를 입력해 주세요."),
     BONUS_NUMBER_REQUEST_MESSAGE("보너스 번호를 입력해 주세요."),
     WINNING_RESULT_NOTICE("당첨 통계" + NEWLINE.getSymbol() + "---"),
     WINNING_RESULT_INFORMATION("%d개 일치 (%,d원) - %d개"),
